Added missing functionality of snowballs
If you like, you could add items to the types of masks that will get the callback which checks if a snowball should be turned into a rock after damaging something, but there's few items where that matters and this call can be expensive. I'm not sure how it would affect performance.