Alienlord adjustments, custom death messages
Made the alienlord more accurate to HD and added the custom death message for him. Also changed the way the squishy alien sounds are played to be less buggy and more accurate to actual gameplay. Also reduced the attack_timer initially, since it only decreases when the player is near there's a long stretch of time where it does nothing even while its in range when you first get near the alienlord.