Surface assets!
might go about updating these in the future, but for now, these are stuffs to be added to the surface! featuring both day and night variants along with palm trees with foreground and background variants..
base_surface2 is the day variant, with duat 2 being the night variant for the special endings.
sky should remove the stars except for the night variant.